Rearrange data functions for more compact package files
